+++ tests/Unit/Api/TimeEntry/ListTest.php
@@ -1,0 +1,77 @@
+<?php
+
+namespace Redmine\Tests\Unit\Api\TimeEntry;
+
+use PHPUnit\Framework\TestCase;
+use Redmine\Api\TimeEntry;
+use Redmine\Client\Client;
+
+/**
+ * @covers \Redmine\Api\TimeEntry::list
+ */
+class ListTest extends TestCase
+{
+    public function testListWithoutParametersReturnsResponse()
+    {
+        // Test values
+        $response = '["API Response"]';
+        $expectedReturn = ['API Response'];
+
+        // Create the used mock objects
+        $client = $this->createMock(Client::class);
+        $client->expects($this->once())
+            ->method('requestGet')
+            ->with('/time_entries.json')
+            ->willReturn(true);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseBody')
+            ->willReturn($response);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseContentType')
+            ->willReturn('application/json');
+
+        // Create the object under test
+        $api = new TimeEntry($client);
+
+        // Perform the tests
+        $this->assertSame($expectedReturn, $api->list());
+    }
+
+    public function testListWithParametersReturnsResponse()
+    {
+        // Test values
+        $parameters = [
+            'project_id' => 5,
+            'user_id' => 10,
+            'limit' => 2,
+        ];
+        $response = '["API Response"]';
+        $expectedReturn = ['API Response'];
+
+        // Create the used mock objects
+        $client = $this->createMock(Client::class);
+        $client->expects($this->once())
+            ->method('requestGet')
+            ->with(
+                $this->logicalAnd(
+                    $this->stringStartsWith('/time_entries.json?'),
+                    $this->stringContains('project_id=5'),
+                    $this->stringContains('user_id=10'),
+                    $this->stringContains('limit=2')
+                )
+            )
+            ->willReturn(true);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseBody')
+            ->willReturn($response);
+        $client->expects($this->exactly(1))
+            ->method('getLastResponseContentType')
+            ->willReturn('application/json');
+
+        // Create the object under test
+        $api = new TimeEntry($client);
+
+        // Perform the tests
+        $this->assertSame($expectedReturn, $api->list($parameters));
+    }
+}